requirements.txt

streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.9.0
pandas>=2.0.0
plotly>=5.0.0
kaleido>=0.2.1
openpyxl>=3.0.0
sqlalchemy>=2.0.0
duckdb>=0.9.0
numpy>=1.24.0